backend/requirements.txt
# Flask Core
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# HTTP & API
requests==2.31.0

# Environment & Config
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

# Optional: Caching
redis==5.0.1

# Development
pytest==7.4.3
pytest-flask==1.3.0
